requirements.txt
streamlit==1.28.0
tensorflow-cpu==2.14.0
opencv-python-headless==4.8.1.78
Pillow==10.0.0
numpy==1.24.3